requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.8.2
click==8.1.7
fastapi==0.115.2
filelock==3.16.1
Flask==3.0.3
fsspec==2024.10.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.2
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
python-multipart==0.0.12
setuptools==75.2.0
sniffio==1.3.1
starlette==0.40.0
sympy==1.13.1
torch==2.5.0
torchvision==0.20.0
typing_extensions==4.12.2
uvicorn==0.32.0
Werkzeug==3.0.4
